2nd September commit

+++ src/main/java/za/ac/cput/service/CustomerService.java
@@ -2,28 +2,42 @@ package za.ac.cput.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import za.ac.cput.domain.Contact;
 import za.ac.cput.domain.Customer;
-import za.ac.cput.repository.ContactRepository;
+import za.ac.cput.dto.CustomerDto;
+import za.ac.cput.factory.CustomerFactory;
 import za.ac.cput.repository.CustomerRepository;
-
+import za.ac.cput.repository.ShippingRepository;
 import java.util.List;
+import java.util.Optional;
 
 @Service
-public class CustomerService implements ICustomerService{
+public class CustomerService implements ICustomerService {
     private CustomerRepository repository;
-    private ContactRepository contactRepository;
+    private ShippingRepository shippingRepository;
 
     @Autowired
-    CustomerService(CustomerRepository repository, ContactRepository contactRepository){
+    CustomerService(CustomerRepository repository, ShippingRepository shippingRepository) {
         this.repository = repository;
-        this.contactRepository = contactRepository;
+        this.shippingRepository = shippingRepository;
     }
+    public void register(CustomerDto customerDto) {
+        Customer customer = CustomerFactory.buildCustomer(
+                customerDto.getFirstName(),
+                customerDto.getLastName(),
+                customerDto.getEmail(),
+                customerDto.getPassword()
+        );
 
+        if (customer == null) {
+            throw new IllegalArgumentException("Invalid customer data");
+        }
+
+        repository.save(customer);
+    }
     @Override
     public Customer create(Customer customer) {
-        contactRepository.save(customer.getContact());
-        return repository.save(customer);}
+        return repository.save(customer);
+    }
 
     @Override
     public Customer read(String id) {
@@ -32,17 +46,26 @@ public class CustomerService implements ICustomerService{
 
     @Override
     public Customer update(Customer customer) {
-        contactRepository.save(customer.getContact());
+        shippingRepository.save(customer.getShipping());
         return repository.save(customer);
     }
 
     @Override
     public void delete(String id) {
-        repository.deleteById(id);
+        if (repository.existsById(id))
+            repository.deleteById(id);
+    }
+
+    public Customer getCustomer(String email){
+        return repository.findByEmail(email);
     }
 
     @Override
     public List<Customer> getAll() {
         return repository.findAll();
+    }
+    public boolean verifyLogin(String email, String password){
+        Optional<Customer> verifyCustomer = repository.findByEmailAndPassword(email, password);
+        return verifyCustomer.isPresent();
     }
 }
